% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcms.R
\docType{data}
\name{continuous_SimpleQ.100}
\alias{continuous_SimpleQ.100}
\title{Continuous DCM (Experimental) Simple Q 100 Data}
\format{A data list of responses, locations etc.}
\usage{
data(continuous_SimpleQ.100)
}
\description{
This is an experimental data set for a continuous DCM
}
\author{
Dave Rackham Oct 2016
}
