% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-HO-R-DINA-JAGS.R
\name{generateHO_R_DINA_Jags}
\alias{generateHO_R_DINA_Jags}
\title{Generate the HO-R-DINA JAGS model}
\usage{
generateHO_R_DINA_Jags()
}
\description{
Generates the HO-R-DINA JAGS model for a simple data set
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\keyword{JAGS}
\keyword{RDINA}
