% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRDINAJags.R
\name{generateRDINAJags}
\alias{generateRDINAJags}
\title{Generate the RDINA Jags model}
\usage{
generateRDINAJags()
}
\description{
Generates the RDINA Jags model for a simple data set
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\keyword{JAGS}
\keyword{RDINA}
