% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRDINAJagsNon-Hierarchical.R
\name{generateRDINAJagsNonHierachical}
\alias{generateRDINAJagsNonHierachical}
\title{Generate the RDINA JAGS Non-Hierachical model}
\usage{
generateRDINAJagsNonHierachical()
}
\description{
Generates the RDINA JAGS Non-Hierachical model for a simple data set
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{JAGS}
\keyword{RDINA}

