% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDADataSims.R
\docType{data}
\name{IRT.1PL}
\alias{IRT.1PL}
\title{IRT 1PL Data}
\format{A data list of responses, locations etc.}
\usage{
data(IRT.1PL)
}
\description{
This data set represents simulated 1PL model data
}
\author{
Dave Rackham May 2016
}

