% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDADataSims.R
\docType{data}
\name{IRT.1PL_100}
\alias{IRT.1PL_100}
\title{IRT 1PL 100 Data}
\format{A data list of responses, locations etc.}
\usage{
data(IRT.1PL_100)
}
\description{
This data set represents simulated 1PL model data for 100 examinees
}
\author{
Dave Rackham May 2016
}

