% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcms.R
\docType{data}
\name{IRT_1PL.1000}
\alias{IRT_1PL.1000}
\title{IRT 1PL 1000 Data}
\format{A data list of responses, locations etc.}
\usage{
data(IRT_1PL.1000)
}
\description{
This data set represents simulated 1PL model data for 1000 examinees
}
\author{
Dave Rackham May 2016
}
