% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDADataSims.R
\docType{data}
\name{IRT_1PL.500}
\alias{IRT_1PL.500}
\title{IRT 1PL 500 Data}
\format{A data list of responses, locations etc.}
\usage{
data(IRT_1PL.500)
}
\description{
This data set represents simulated 1PL model data for 500 examinees
}
\author{
Dave Rackham May 2016
}

