% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcms.R
\docType{data}
\name{R_DINA_SimpleQ.500}
\alias{R_DINA_SimpleQ.500}
\title{R-DINA Simple Q 500 Data}
\format{A data list of responses, locations etc.}
\usage{
data(R_DINA_SimpleQ.500)
}
\description{
This data set represents simulated R-DINA Simple Q for 500 examinees
}
\author{
Dave Rackham May 2016
}

