% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcms.R
\docType{data}
\name{R_DINA_SimpleQ_HN.1000}
\alias{R_DINA_SimpleQ_HN.1000}
\title{R-DINA Simple Q High noise 1000 Data}
\format{A data list of responses, locations etc.}
\usage{
data(R_DINA_SimpleQ_HN.1000)
}
\description{
This data set represents simulated R-DINA Simple Q with high noise
}
\author{
Dave Rackham June 2016
}
