% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifiedRumSim.R
\name{simplifiedRUMSimSTAN}
\alias{simplifiedRUMSimSTAN}
\title{Simplified RUM Simulation in STAN}
\usage{
simplifiedRUMSimSTAN()
}
\description{
Fit a simplified RUM model to simulated data and evaluate data-model fit.
}
\examples{
simplifiedRUMSimSTAN()
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{RUM}
\keyword{STAN}
\keyword{hartz}
\keyword{roussos}

