% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAN-IRT-1PL.R
\name{STAN_IRT_1PL}
\alias{STAN_IRT_1PL}
\title{IRT 1PL  Simulation in STAN}
\usage{
STAN_IRT_1PL(model = "IRT-1PL.stan", data, wd, cores, iter, chains)
}
\arguments{
\item{model}{Model to be used. Must be available in working directory}

\item{data}{Dataset to be used}

\item{wd}{Desired working directory (string)}

\item{cores}{Number of cores to use}

\item{iter}{Number of iterations}

\item{chains}{Number of chains}
}
\description{
Fit a 1PL model to simulated data using STAN
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\keyword{1PL}
\keyword{IRT}
\keyword{STAN}

