% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDINAStan.R
\name{stanSim}
\alias{stanSim}
\title{R-DINA Simulation in STAN}
\usage{
stanSim(model = "R-DINA-Non-Logit.stan", data, wd, cores, iter, chains)
}
\arguments{
\item{model}{Model to use}

\item{data}{Dataset to be used}

\item{wd}{Desired working directory (string)}

\item{cores}{Number of cores to use}

\item{iter}{Number of iterations}

\item{chains}{Number of chains}
}
\description{
Fit a R-DINA model to simulated data using STAN
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\keyword{R-DINA}
\keyword{STAN}

