% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateExplicitPriorsRUMJags.R
\name{generateExplicitPriorsRUMJags}
\alias{generateExplicitPriorsRUMJags}
\title{Generate the Explicit Priors RUM Jags model}
\usage{
generateExplicitPriorsRUMJags(I, J, K, q)
}
\description{
Generate the Explicit Priors RUM Jags model
}

