% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifiedRUMJags.R
\name{runJagsSim}
\alias{runJagsSim}
\title{Simplified RUM Simulation in JAGS}
\usage{
runJagsSim(data, q = NULL, jagsModel = "simplifiedRUM.jags",
  adaptSteps = 10, burnInSteps = 40, numSavedSteps = 10, thinSteps = 1,
  dynamicModel = TRUE)
}
\arguments{
\item{data}{Uses the dcmdata simplifiedRUM function to generate data}

\item{q}{Q-matrix}

\item{jagsModel}{The specified model for JAGS to use}

\item{adaptSteps}{Adapt Steps}

\item{burnInSteps}{Burn-in Steps}

\item{numSavedSteps}{Saved Steps}

\item{thinSteps}{Thin Steps}

\item{dynamicModel}{Logical that specifies if runJagsSim should look for the model within the package}
}
\description{
Fit a simplified RUM model to simulated data and evaluate data-model fit using JAGS.
}
\author{
Dave Rackham \email{ddrackham@gmail.com}
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1002/j.2333-8504.2008.tb02157.x/abstract}
}
\keyword{JAGS}
\keyword{RUM}
\keyword{hartz}
\keyword{roussos}

